Revert "moving tests from page.spec.tsx"
This reverts commit 7ab12382925637d82224d4908ac1a905f0b32a2b.
